moved JSONPointer and DefaultSchemaClient classes to org.everit.json.schema.loader.internal package, hence now thy are not parts of the public API of the bundle
