+++ core/src/test/java/org/everit/json/schema/loader/SchemaLoaderTest.java
@@ -32,9 +32,7 @@ import org.everit.json.schema.ReferenceSchema;
 import org.everit.json.schema.Schema;
 import org.everit.json.schema.SchemaException;
 import org.everit.json.schema.StringSchema;
-import org.everit.json.schema.loader.DefaultSchemaClient;
-import org.everit.json.schema.loader.SchemaClient;
-import org.everit.json.schema.loader.SchemaLoader;
+import org.everit.json.schema.loader.internal.DefaultSchemaClient;
 import org.json.JSONArray;
 import org.json.JSONObject;
 import org.json.JSONTokener;
